+++ Transactions_microService/package.json
@@ -25,6 +25,7 @@
     "@nestjs/core": "^8.0.0",
     "@nestjs/mongoose": "^8.0.0",
     "@nestjs/platform-express": "^8.0.0",
+    "@nestjs/typeorm": "^8.0.2",
     "mongoose": "^5.13.7",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
