--- Transactions_microService/src/account/account.model.ts
@@ -1,31 +1,0 @@
-import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
-import { Document } from 'mongoose';
-
-export type AccountDocument = Account & Document;
-
-export type balanceType = {
-  amount: number;
-  history: Array<transactionType>;
-};
-
-export type transactionType = {
-  senderEmail: string;
-  receiverEmail: string;
-  type: string;
-  value: number;
-  date: Date;
-};
-
-@Schema()
-export class Account {
-  @Prop({ required: true, unique: true })
-  email: string;
-
-  @Prop({ required: true, type: Object })
-  balance: balanceType;
-
-  @Prop({ required: true })
-  cvu: string;
-}
-
-export const AccountSchema = SchemaFactory.createForClass(Account);
