+++ Transactions_microService/src/account/account.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { AccountService } from './account.service';
 import { AccountController } from './account.controller';
 import { MongooseModule } from '@nestjs/mongoose';
-import { AccountSchema } from './account.model';
+import { AccountSchema } from './account.entity';
 
 @Module({
   imports: [
