+++ Transactions_microService/src/account/account.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Account, AccountDocument, transactionType } from './account.model';
+import { Account, AccountDocument } from './account.entity';
 import { Model } from 'mongoose';
 // import { Connection } from 'typeorm';
 
@@ -41,49 +41,49 @@ export class AccountService {
     }
   }
 
-  async updateAccount(tran: transactionType) {
-    // const queryRunner = this.connection.createQueryRunner();
-    // await queryRunner.connect();
-    // await queryRunner.startTransaction();
-    // try {
-    //   if (tran.receiverEmail === tran.senderEmail) {
-    //     const findSender = await this.accountModel.findOne({
-    //       email: tran.senderEmail,
-    //     });
-    //     findSender.balance.amount += tran.value;
-    //     findSender.balance.history.push(tran);
-    //     await queryRunner.manager.update(
-    //       this.accountModel,
-    //       tran.senderEmail,
-    //       findSender,
-    //     );
-    //   } else {
-    //     const findSender = await this.accountModel.findOne({
-    //       email: tran.senderEmail,
-    //     });
-    //     findSender.balance.amount -= tran.value;
-    //     findSender.balance.history.push(tran);
-    //     await queryRunner.manager.update(
-    //       this.accountModel,
-    //       tran.senderEmail,
-    //       findSender,
-    //     );
-    //     const findReceiver = await this.accountModel.findOne({
-    //       email: tran.receiverEmail,
-    //     });
-    //     findReceiver.balance.amount += tran.value;
-    //     findReceiver.balance.history.push(tran);
-    //     await queryRunner.manager.update(
-    //       this.accountModel,
-    //       tran.receiverEmail,
-    //       findReceiver,
-    //     );
-    //   }
-    // } catch (err) {
-    //   console.log(err);
-    //   await queryRunner.rollbackTransaction();
-    // } finally {
-    //   await queryRunner.release();
-    // }
-  }
+  // async updateAccount(tran: transactionType) {
+  // const queryRunner = this.connection.createQueryRunner();
+  // await queryRunner.connect();
+  // await queryRunner.startTransaction();
+  // try {
+  //   if (tran.receiverEmail === tran.senderEmail) {
+  //     const findSender = await this.accountModel.findOne({
+  //       email: tran.senderEmail,
+  //     });
+  //     findSender.balance.amount += tran.value;
+  //     findSender.balance.history.push(tran);
+  //     await queryRunner.manager.update(
+  //       this.accountModel,
+  //       tran.senderEmail,
+  //       findSender,
+  //     );
+  //   } else {
+  //     const findSender = await this.accountModel.findOne({
+  //       email: tran.senderEmail,
+  //     });
+  //     findSender.balance.amount -= tran.value;
+  //     findSender.balance.history.push(tran);
+  //     await queryRunner.manager.update(
+  //       this.accountModel,
+  //       tran.senderEmail,
+  //       findSender,
+  //     );
+  //     const findReceiver = await this.accountModel.findOne({
+  //       email: tran.receiverEmail,
+  //     });
+  //     findReceiver.balance.amount += tran.value;
+  //     findReceiver.balance.history.push(tran);
+  //     await queryRunner.manager.update(
+  //       this.accountModel,
+  //       tran.receiverEmail,
+  //       findReceiver,
+  //     );
+  //   }
+  // } catch (err) {
+  //   console.log(err);
+  //   await queryRunner.rollbackTransaction();
+  // } finally {
+  //   await queryRunner.release();
+  // }
+  // }
 }
