+++ Transactions_microService/src/app.module.ts
@@ -1,14 +1,20 @@
 import { Module } from '@nestjs/common';
-import { MongooseModule } from '@nestjs/mongoose';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
+import { TypeOrmModule } from '@nestjs/typeorm';
 import { TransactionsModule } from './transactions/transactions.module';
+import { Transaction } from './transactions/transactions.entity';
 
 @Module({
   imports: [
     TransactionsModule,
-    MongooseModule.forRoot('mongodb://localhost/banco-del-sol', {
-      autoCreate: true,
+    TypeOrmModule.forRoot({
+      type: 'mongodb',
+      host: 'localhost',
+      database: 'banco-del-sol',
+      entities: [Transaction],
+      synchronize: true,
+      useUnifiedTopology: true,
     }),
   ],
   controllers: [AppController],
