+++ Transactions_microService/src/transactions/transactions.controller.ts
@@ -1,23 +1,9 @@
-import {
-  Controller,
-  Get,
-  Res,
-  HttpStatus,
-  Put,
-  Body,
-  UseGuards,
-  Req,
-  Post,
-} from '@nestjs/common';
+import { Controller, Get, Res, HttpStatus, Body, Post } from '@nestjs/common';
 import { TransactionsService } from './transactions.service';
-import { AccountService } from '../account/account.service';
 
 @Controller('transactions')
 export class TransactionsController {
-  constructor(
-    private transactionService: TransactionsService,
-    private accountService: AccountService,
-  ) {}
+  constructor(private transactionService: TransactionsService) {}
 
   @Get()
   async getTransactions(@Res() res, @Body() body) {
@@ -36,7 +22,7 @@ export class TransactionsController {
   async newTransaction(@Res() res, @Body() body) {
     try {
       const succeeded = await this.transactionService.createTransaction(body);
-      return res.status(HttpStatus.OK).send(succeeded);
+      return res.send(succeeded);
     } catch (error) {
       console.error(error);
       return res.status(400);
